% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack - brute force algorithm}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{as data frame}

\item{W}{as numeric}
}
\value{
value_knap
}
\description{
Knapsack - brute force algorithm
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
\author{
Brian Masinde, Maria Treesa Sebastian, Omkar
}
