% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackdynamic.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{knapsack - Dynamic Algorithm}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{as dataframe}

\item{W}{as numeric}
}
\value{
list containing value and elements
}
\description{
knapsack - Dynamic Algorithm
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1: 8,], W = 3500)


}
\author{
Maria,Masinde,Omkar
}
