% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Function to solve knapsack problem using dynammic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{as data.frame of weights and profits}

\item{W}{as maximum kapsack weight}
}
\value{
a list
}
\description{
Function to solve knapsack problem using dynammic programming
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
}
