% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack Objects Data}
\format{A data frame with 2000 rows and 2 variables
\describe{
  \item{w}{weight of the items}
  \item{v}{value of the items}
}}
\usage{
knapsack_objects
}
\description{
Knapsack Objects Data
}
\keyword{datasets}
