% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack Solution for Kanpsack Problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a dataframe constising}

\item{W}{is capacity of kanpsack}
}
\value{
\code{list} List of object containing \code{value} giving maximum value of Knapsack out of dataframe and \code{elements} giving postion of 
selected values from dataframe x
}
\description{
Approach is to use the a heuristic or approximation for the problem. This algorithm will not
give an exact result :but it can be shown that it will return at least 50percent of the true maximum value:,
but it will reduce the computational complexity considerably :actually to Oh n log n: due to the sorting
part of the algorithm.
}
\examples{
\dontrun{
greedy_knapsack(data.frame("w" = c(10,20,30,40) , "v" = c(2,4,3,5)), 3500)
 }
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
