% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Function to solve knapsack problem using greedy algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{as data.frame of weights and profits}

\item{W}{as maximum kapsack weight}
}
\value{
a list
}
\description{
Function to solve knapsack problem using greedy algorithm
}
